+++ openmc/__init__.py
@@ -30,7 +30,7 @@ from openmc.mixin import *
 from openmc.plotter import *
 from openmc.search import *
 from openmc.polynomial import *
-from openmc.trackfile import *
+from openmc.tracks import *
 from . import examples
 
 # Import a few names from the model module
